config: split the WireGuard unit-device rule out of the interface compiler
Adding the #6941 rule and its rationale pushed compiler_interfaces.go
from 1494 to 1553 LOC, past the 1500 LOC [WATCH] modularity floor. Split
rather than exempted: the rule is a self-contained question about ONE
leaf -- which Linux device a unit's tunnel stanza resolves to -- so it
reads better beside its own reasoning than inline in the interface
compiler. compiler_interfaces.go is back to 1498.

Also fixes the pkg/configstore sibling of the #6861 fixture edit.
TestSyncApplyRendersThePointerKeyedAnchor_4785 builds the same
Name/device divergence with `unit 1 tunnel mode wireguard`, and needs it
for the same reason: the divergence keeps gr-0-0-0 out of the live
device set so the anchor advisory fires on the dead ipip record. Its
lever moves to `mode gre` like its pkg/config twin, with the reason
recorded at the fixture.

That one was missed by the first matrix, which ran only pkg/config and
pkg/dataplane/userspace. It surfaced on the repo-wide gate, which is
what the repo-wide gate is for.
